splitting exstream.test in more logical tests, now we have one for fs (slow) and one generic slow
