+++ test/exstream.test.js
@@ -1,18 +1,9 @@
 /* eslint-disable max-lines */
-jest.setTimeout(2000)
-
 const _ = require('../src/index.js')
 const h = require('./helpers.js')
 const EventEmitter = require('events').EventEmitter
-const zlib = require('zlib')
 jest.mock('fs')
 const fs = require('fs')
-
-const out = [...h.randomStringGenerator(10000)].map(x => x.toString() + '\n')
-
-beforeEach(() => {
-  fs.__setMockFiles({ out })
-})
 
 test('stream initialization', () => {
   const x = _([1, 2, 3])
@@ -44,28 +35,6 @@ test('consume stream', () => {
     expect(y).toEqual([1, 2, 3])
   })
   x.resume()
-})
-
-test('backpressure', () => {
-  const x = _([1, 2, 3])
-  const y = []
-  return new Promise(resolve => {
-    const z = x.consume((err, x, push, next) => {
-      if (err) {
-        y.push(err)
-        setTimeout(() => next(), 100)
-      } else if (x !== _.nil) {
-        y.push(x)
-        setTimeout(() => next(), 100)
-      } else {
-        push(null, _.nil)
-      }
-    }).on('end', () => {
-      expect(y).toEqual([1, 2, 3])
-      resolve()
-    })
-    z.resume()
-  })
 })
 
 test('write', () => {
@@ -359,17 +328,6 @@ test('reject', () => {
   expect(res).toEqual([1, 3])
 })
 
-test('async filter', async () => {
-  const res = await _([1, 2, 3])
-    .asyncFilter(async x => {
-      await h.sleep(100)
-      return x % 2 === 0
-    })
-    .toPromise()
-
-  expect(res).toEqual([2])
-})
-
 test('through pipeline', () => {
   _([1, 2, 3])
     .through(_.pipeline()
@@ -422,59 +380,6 @@ test('toPromise', async () => {
     .toPromise()
 
   expect(res).toEqual([2, 4, 6])
-})
-
-test('unordered promises', async () => {
-  let sleepCount = 3
-  const sleep = x => new Promise(resolve => setTimeout(() => resolve(x), 200 * sleepCount--))
-
-  const res = await _([2, 3, 4])
-    .map(x => sleep(x))
-    .massThen(x => x * 2)
-    .massThen(x => x * 2)
-    .resolve(2, false)
-    .toPromise()
-
-  expect(res).toEqual([12, 8, 16])
-})
-
-test('promises hl style', async () => {
-  let sleepCount = 3
-  const sleep = x => new Promise(resolve => setTimeout(() => resolve(x), 200 * sleepCount--))
-
-  const res = await _([2, 3, 4])
-    .map(x => _(sleep(x)))
-    .merge(3, true)
-    .toPromise()
-
-  expect(res).toEqual([2, 3, 4])
-})
-
-test('ordered promises', async () => {
-  let sleepCount = 3
-  const decrementalSlowMap = x => new Promise(resolve => setTimeout(() => resolve(x), 200 * sleepCount--))
-
-  const res = await _([2, 3, 4])
-    .map(x => decrementalSlowMap(x))
-    .massThen(x => x * 2)
-    .massThen(x => x * 2)
-    .resolve(3)
-    .toPromise()
-
-  expect(res).toEqual([8, 12, 16])
-})
-
-test('slow writes on node stream', () => {
-  const res = []
-  return new Promise(resolve => {
-    _([2, 3, 4])
-      .map(x => x * 2)
-      .pipe(h.getSlowWritable(res))
-      .on('finish', () => {
-        resolve()
-        expect(res).toEqual([4, 6, 8])
-      })
-  })
 })
 
 test('promise in constructor', () => {
@@ -584,22 +489,6 @@ test('switch source + backpressure', done => {
   })
 })
 
-test('throttle', async () => {
-  const gen = async function * () {
-    for (let i = 0; i < 10; i++) {
-      await h.sleep(8)
-      yield i
-    }
-  }
-
-  const res = await _(gen())
-    .throttle(50)
-    .toPromise()
-
-  expect(res.length).toBeLessThan(4)
-  expect(res.length).toBeGreaterThan(1)
-})
-
 test('split', () => {
   const b = [Buffer.from('line1\nli'), Buffer.from('ne2\r\n'), Buffer.from('line3')]
   const res = _(b).split().values()
@@ -649,23 +538,6 @@ test('toNodeStream', () => {
   })
 })
 
-test('through node stream', () => new Promise(resolve => {
-  _(fs.createReadStream('out'))
-    .through(zlib.createGzip())
-    .pipe(fs.createWriteStream('out.gz'))
-    .on('finish', () => {
-      _(fs.createReadStream('out.gz'))
-        .through(zlib.createGunzip())
-        .pipe(fs.createWriteStream('out2'))
-        .on('finish', () => {
-          const f1 = fs.readFileSync('out')
-          const f2 = fs.readFileSync('out2')
-          expect(f1).toEqual(f2)
-          resolve()
-        })
-    })
-}))
-
 test('forking', async () => {
   const s = _([1, 2, 3])
   const p1 = s.fork().map(x => x * 2 + 1).toPromise()
@@ -675,38 +547,6 @@ test('forking', async () => {
   expect(r1).toEqual([3, 5, 7])
   expect(r2).toEqual([4, 6, 8])
   expect(r3).toEqual([5, 7, 9])
-})
-
-test('merging with fs', async () => new Promise(resolve => {
-  _([
-    _(fs.createReadStream('out')),
-    _(fs.createReadStream('out')),
-  ]).merge(1)
-    .pipe(fs.createWriteStream('out3'))
-    .on('finish', () => {
-      resolve()
-      const o = out.map(x => x.toString()).join('')
-      expect(fs.__getMockFiles().out3.join('')).toEqual(o + o)
-    })
-}))
-
-test('pipe pipeline', done => {
-  const p = _.pipeline()
-    .map(x => x.toString())
-    .take(10)
-    .collect()
-    .map(x => x.join().split('\n'))
-    .flatten()
-    .map(x => 'buahaha' + x + '\n')
-
-  const res = []
-  fs.createReadStream('out')
-    .pipe(p.generateStream())
-    .pipe(h.getSlowWritable(res, 0))
-    .on('finish', () => {
-      done()
-      expect(res.length).toBe(11)
-    })
 })
 
 test('pipeToFile', () => new Promise(resolve => {
