+++ test/recursive.test.js
@@ -2,7 +2,7 @@ const _ = require('../src/index.js')
 const h = require('./helpers')
 
 const increment = ({ begin, end }) => _((push, next) => {
-  console.log({ begin, end })
+  // console.log({ begin, end })
   const current = begin + 1
   if (current < end) {
     push(current)
@@ -13,7 +13,6 @@ const increment = ({ begin, end }) => _((push, next) => {
 })
 
 const incrementAsync = ({ begin, end }) => _(async (push, next) => {
-  console.log({ begin, end })
   await h.sleep(1)
   const current = begin + 1
   if (current < end) {
@@ -27,14 +26,13 @@ const incrementAsync = ({ begin, end }) => _(async (push, next) => {
 
 test('iterate', async () => {
   const values = await _(increment({ begin: 0, end: 10 }))
-    .tap(console.log)
+    // .tap(console.log)
     .toPromise()
   expect(values).toEqual([1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
 })
 
 test('iterateAsync', async () => {
   const values = await _(incrementAsync({ begin: 0, end: 10 }))
-    .tap(console.log)
     .on('end', () => console.log('ended'))
     .toPromise()
   expect(values).toEqual([1, 2, 3, 4, 5, 6, 7, 8, 9, 10])
